fix(axbuild): isolate host test process state
